+++ cpp06/ex01/Serializer.hpp
@@ -6,7 +6,7 @@
 /*   By: pgrellie <pgrellie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/02 15:04:50 by pgrellie          #+#    #+#             */
-/*   Updated: 2025/07/02 18:21:55 by pgrellie         ###   ########.fr       */
+/*   Updated: 2025/07/07 17:54:47 by pgrellie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 
 # include <string>
 # include <iostream>
+# include <stdint.h>
 
 struct Data
 {
@@ -26,8 +27,8 @@ class	Serializer{
 
 public:
 
-	static unsigned long	serialize(Data *ptr);
-	static Data				*deserialize(unsigned long raw);
+	static uintptr_t		serialize(Data *ptr);
+	static Data				*deserialize(uintptr_t raw);
 
 private:
